requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
Django==4.2.6
django-bootstrap-v5==1.0.11
django-ckeditor==6.7.0
django-excel-response2==3.0.6
django-js-asset==2.2.0
django-six==1.0.5
excel-base==1.0.4
isoweek==1.3.3
Pillow==10.1.0
python-dateutil==2.8.2
screen==1.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
TimeConvert==3.0.13
typing-extensions==4.8.0
tzlocal==5.1
xlwt==1.3.0
